+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 
 🔥 *Features*
 * Force stop workflow runs via the CLI or Python API
+* `WorkflowRun.get_tasks()` supports filtering tasks by state, function name, task run ID and task invocation ID.
 
 👩‍🔬 *Experimental*
 
@@ -23,6 +24,7 @@
 
 🚨 *Breaking Changes*
 * `WorkflowRun.get_logs()` now returns a data structure with logs split into categories.
+* Local workflow database has changed format: downgrading from v0.50.0 will require user intervention.
 
 🔥 *Features*
 * Add .project property to WorkflowRun to get the info about workspace and project of running workflow